segregate pymathic and a pattern  test
