+++ mathics/builtin/patterns.py
@@ -1658,29 +1658,6 @@ class OptionsPattern(PatternObject):
         return [element for element in elements if _match(element)]
 
 
-class _StopGeneratorBaseElementIsFree(StopGenerator):
-    pass
-
-
-def item_is_free(item, form, evaluation):
-    # for vars, rest in form.match(self, {}, evaluation, fully=False):
-    def yield_match(vars, rest):
-        raise _StopGeneratorBaseElementIsFree(False)
-        # return False
-
-    try:
-        form.match(yield_match, item, {}, evaluation, fully=False)
-    except _StopGeneratorBaseElementIsFree as exc:
-        return exc.value
-
-    if isinstance(item, Atom):
-        return True
-    else:
-        return item_is_free(item.head, form, evaluation) and all(
-            item_is_free(element, form, evaluation) for element in item.elements
-        )
-
-
 class Dispatch(Atom):
     class_head_name = "System`Dispatch"
